+++ hooli_data_eng/__init__.py
@@ -1,1 +1,1 @@
-from .definitions import defs as defs
+from .definitions import defs
