+++ hooli_data_eng/databricks/definitions.py
@@ -1,0 +1,30 @@
+from hooli_data_eng.databricks import assets as databricks_assets
+from hooli_data_eng.databricks.resources import resource_def
+from hooli_data_eng.utils import get_env
+
+from pathlib import Path
+import dagster as dg
+from dagster_cloud.metadata.source_code import link_code_references_to_git_if_cloud
+
+# Our first set of assets represent raw data, and the asset
+# definitions can be seen in /assets/raw_data/__init__.py
+#
+# These raw datasets will be used by dbt as dbt sources
+# which can be found in dbt_project/models/sources.yml
+databricks_assets = dg.load_assets_from_modules(
+    [databricks_assets], group_name="FORECASTING"
+)
+
+defs = dg.Definitions(
+        assets=link_code_references_to_git_if_cloud(
+            dg.with_source_code_references(
+                [*databricks_assets,
+                ]
+            ),
+            file_path_mapping=dg.AnchorBasedFilePathMapping(
+                local_file_anchor=Path(__file__),
+                file_anchor_path_in_repository="hooli_data_eng/databricks/definitions.py",
+            ),
+        ),
+        resources=resource_def[get_env()],
+)
