+++ hooli_data_eng/definitions.py
@@ -1,71 +1,29 @@
-from pathlib import Path
-from dagster import (
-    AnchorBasedFilePathMapping,
-    Definitions,
-    load_assets_from_modules,
-    load_assets_from_package_module,
-    build_column_schema_change_checks,
-    multiprocess_executor,
-    with_source_code_references,
-)
-from dagster.components import load_defs
-from dagster_cloud.metadata.source_code import link_code_references_to_git_if_cloud
+import warnings 
 
+from dagster._utils.warnings import BetaWarning, PreviewWarning
+warnings.filterwarnings("ignore", category=PreviewWarning)
+warnings.filterwarnings("ignore", category=BetaWarning)
+
+import dagster as dg
+from dagster.components import load_defs
+
+import hooli_data_eng.dbt.definitions as dbt_definitions
+import hooli_data_eng.custom_ingest.definitions as custom_ingest_definitions
+import hooli_data_eng.databricks.definitions as databricks_definitions
+import hooli_data_eng.kubernetes.definitions as kubernetes_definitions
+import hooli_data_eng.scikit_learn.definitions as scikit_learn_definitions
+import hooli_data_eng.files.definitions as files_definitions
+import hooli_data_eng.pandas.definitions as pandas_definitions
+import hooli_data_eng.notebooks.definitions as notebooks_definitions
+import hooli_data_eng.spark.definitions as spark_definitions   
 import hooli_data_eng.defs
-from hooli_data_eng.assets import forecasting, raw_data, marketing, dbt_assets
-from hooli_data_eng.assets.dbt_assets import dbt_slim_ci_job
-from hooli_data_eng.assets.marketing import check_avg_orders
-from hooli_data_eng.assets.raw_data import check_users, raw_data_schema_checks
-from hooli_data_eng.jobs import analytics_job, predict_job
-from hooli_data_eng.resources import get_env, resource_def
-from hooli_data_eng.schedules import analytics_schedule
-from hooli_data_eng.sensors import orders_sensor, dbt_code_version_sensor
-from hooli_data_eng.sensors.watch_s3 import watch_s3_sensor
-from hooli_data_eng.assets.marketing import (
-    avg_orders_freshness_check,
-    min_order_freshness_check,
-    min_order_freshness_check_sensor,
-    avg_orders_freshness_check_schedule,
-)
-from hooli_data_eng.assets.dbt_assets import (
-    weekly_freshness_check,
-    weekly_freshness_check_sensor,
-)
+
+
 # ---------------------------------------------------
 # Assets
 
 # Dagster assets specify what outputs we care about and
 # the logical code needed to create them
-
-# Our first set of assets represent raw data, and the asset
-# definitions can be seen in /assets/raw_data/__init__.py
-#
-# These raw datasets will be used by dbt as dbt sources
-# which can be found in dbt_project/models/sources.yml
-raw_data_assets = load_assets_from_package_module(
-    raw_data, group_name="RAW_DATA", key_prefix=["RAW_DATA"]
-)
-
-# Our second set of assets represent dbt models
-# these models are defined in the dbt_project
-# folder
-#
-# The dbt file dbt_project/config/profiles.yaml
-# specifies what databases to targets, and locally will
-# execute against a DuckDB
-
-dbt_assets = load_assets_from_modules([dbt_assets])
-
-dbt_asset_checks = build_column_schema_change_checks(assets=[*dbt_assets])
-
-# Our final set of assets represent Python code that
-# should run after dbt. These assets are defined in
-# assets/forecasting/__init__.py
-forecasting_assets = load_assets_from_package_module(
-    forecasting, group_name="FORECASTING"
-)
-
-marketing_assets = load_assets_from_package_module(marketing, group_name="MARKETING")
 
 
 # ---------------------------------------------------
@@ -74,43 +32,20 @@ marketing_assets = load_assets_from_package_module(marketing, group_name="MARKET
 # Definitions are the collection of assets, jobs, schedules, resources, and sensors
 # used with a project. Dagster Cloud deployments can contain mulitple projects.
 
-defs = Definitions.merge(
+
+defs = dg.Definitions.merge(
     load_defs(hooli_data_eng.defs),
-    Definitions(
-        executor=multiprocess_executor.configured({"max_concurrent": 3}),
-        assets=link_code_references_to_git_if_cloud(
-            with_source_code_references(
-                [
-                    *dbt_assets,
-                    *raw_data_assets,
-                    *forecasting_assets,
-                    *marketing_assets,
-                ]
-            ),
-            file_path_mapping=AnchorBasedFilePathMapping(
-                local_file_anchor=Path(__file__),
-                file_anchor_path_in_repository="hooli_data_eng/definitions.py",
-            ),
-        ),
-        asset_checks=[
-            *raw_data_schema_checks,
-            *dbt_asset_checks,
-            check_users,
-            check_avg_orders,
-            *min_order_freshness_check,
-            *avg_orders_freshness_check,
-            *weekly_freshness_check,
-        ],
-        resources=resource_def[get_env()],
-        schedules=[analytics_schedule, avg_orders_freshness_check_schedule],
-        sensors=[
-            orders_sensor,
-            watch_s3_sensor,
-            #       asset_delay_alert_sensor,
-            min_order_freshness_check_sensor,
-            dbt_code_version_sensor,
-            weekly_freshness_check_sensor,
-        ],
-        jobs=[analytics_job, predict_job, dbt_slim_ci_job],
+    dbt_definitions.defs,
+    custom_ingest_definitions.defs,
+    databricks_definitions.defs,
+    scikit_learn_definitions.defs,
+    kubernetes_definitions.defs,
+    files_definitions.defs,
+    pandas_definitions.defs,
+    notebooks_definitions.defs,
+    spark_definitions.defs,
+    dg.Definitions(
+        # only apply this setting once
+        executor=dg.multiprocess_executor.configured({"max_concurrent": 3}),
     )
 )
